Teste 8 falhando (desvioPadrão)

+++ src/main/java/com/utfpr/classes/CalculaMediana.java
@@ -23,13 +23,13 @@ public class CalculaMediana {
             if(vetor.length % 2 == 0){
                 int posicao1 = (int) ((vetor.length / 2) - 0.5f);
                 int posicao2 = (int) ((vetor.length / 2) + 0.5f);
-                res = (double) ((vetor[posicao1]+vetor[posicao2])/2);
+                res = (float) ((vetor[posicao1]+vetor[posicao2])/2);
                 return res;
             }else{
                 
                 if(vetor.length % 2 != 0){
                     int posicao = vetor.length / 2;
-                    res =(double) vetor[posicao];
+                    res =(float) vetor[posicao];
                     return res;
                 }
             }
